Primitives implementiert, das dloading ist erstmal emuliert, ist unter c++ etwas komplizierter... richtige implementierung des primitiven laden kommt noch gettimeofday gibt es scheinbar in windows nicht

+++ src/vmobjects/VMObject.cpp
@@ -9,7 +9,7 @@ VMObject::VMObject()
 VMObject::VMObject( int number_of_fields )
 {
     gcfield = 0; 
-	hash = (uint16_t)this;
+	hash = (int32_t)this;
 	objectSize = sizeof(VMObject) + number_of_fields*sizeof(VMObject*);
     numberOfFields = number_of_fields;
     fields = (VMObject**)&this->clazz;
